+++ BE/src/main/java/sidedish/SidedishApplication.java
@@ -1,13 +1,13 @@
-package be.demo;
+package sidedish;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class SidedishApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
+        SpringApplication.run(SidedishApplication.class, args);
     }
 
 }
