+++ BE/src/main/java/sidedish/domain/Category.java
@@ -1,0 +1,4 @@
+package sidedish.domain;
+
+public class Category {
+}
